golf/3Dcam/front_realsense/requirements.txt
requests
pyjwt
pyrealsense2
numpy
opencv-python
pyinstaller
python-dotenv
